backend/requirements_rasa.txt
pip-tools
Flask
flask-cors
SpeechRecognition
gTTS
pydub
huggingface_hub
torch
numpy==1.23.5
rasa==3.6.21
tensorflow-intel==2.12.0
vosk
pyaudio
transformers accelerate
python-dotenv
pipwin
pyttsx3
